Bug Fixes, Dummy Methods & A Compiled Build
Fixed the way the dictionaries are loaded into the binary tree and
changed the way we order the binary tree. This was changed to using a
either the french or english word as the ID and using the compareTo()
method to order the nodes in the tree. Some bugs were in Momokos code so
I fixed them while I was adding in this new loading feature. I also made
some dummy methods for the the stuff that needs to be called for the
program to compile ann executable jar for program to test how long it
would take to load everything in. (Turns out around 8 mins 😱 but I'm
working on that so don't worry)