Updates and Encapsulation
Fixed some of the code for displaying details in the DictionaryTree
class so we didn't need to add another method to the WordNode class.
Also made it so you don't need to reference a binary tree within the
menu class. All that is now handled wtihin the Translator class with a
brand new displayDictionary() method. (Also rebuilt the runnable jar for
testing purposes)