Fixed the Loading Time
No longer takes 8 mins to load instead takes around 10 secs! 😂 The
wonders of inputting your data randomly are truly incredible. Made a new
jar testing build for those of you who want to try out the methods
already implemented and check for bugs. (Loading is slightly faster on
the jar but 10 secs is 10 secs yah know) Also the new dictionary file is
a combination of the other two strictly used for inputing the data
randomly into the binary tree.