Added Mutli Word French to English Word Translation
Also the previous commit was a lie, a terrible terrible lie. Turns out
the deleteWords() method has a bug that makes it not delete the french
translation when it deletes the english one. So try to avoid using that
in testing for now. This new commit brings slightly less lies with multi
word french to single word english translation as well as a new jar file
build to test things on.